Moved out some more shared code and did a bit more cleanup.
